divide: Move data formatting out of decode2
This moves the data formatting (sign extension of 32-bit operands, and
shifting of the dividend for 32-bit extended divide operations) out of
decode2 into the first cycle of the divider pipe.  This will help make
it possible to do forwarding of results from the execute pipe at the
front of execute1.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>